Filter out data from stripe before passing it client-side
Specifically, hash any ids that we pass through with the ROOT_URL used
as a salt. These ids probably aren't all that important since you'd need
access to the stripe key to do anything with them anyways (which would
grant enumeration power anyways), but we might as well be safe with it.