Fix bug in storage causing crashes.
The crash occurred when a Volume was modified after its parent object had been deleted. An attempt to update the transitive size counter on the parent would throw an exception because the parent file no longer existed. We don't actually need to update anything in this case, so the crash was a false alarm.