Implement storage bonus for people on our mailing list.
If any of the user's verified addresses are subscribed, we give them a bonus.

This change does NOT yet add a way to one-click subscribe or unsubscribe from inside the UI, but that is intended to come next.

This change also implements "gift from the admin": bonuses applied via Stripe metadata (editable in the Stripe console). We have already assigned certain Indiegogo bonuses this way, so they will now be honored.