Fix race condition when multiple simultaneous attepmts are made to start the same grain.
This caused the "assignable modified concurrently" exception observed a few times in production.

I suspect this also caused the problem where a grain would get stuck in a state where it appears never to load in the browser, even though the grain is in fact running happily and writing to its log. I suspect this problem would come about in cases where the front-end was stuck trying to communicate with a different instance of the grain that was locked up after its storage had been pulled out from under it due to this race. I am not sure about this, though.