Compile and bundle e2fsprogs.
The e2fsprogs installed in prod are too old to understand ext4 volumes that have been built with new vensions of mke2fs, which many grains in production were. We need to bundle new versions, and statically-link them because the libs in production are also old.

Additionally, we need to compile e2fsck with -DEXT2_SKIP_UUID in order to prevent it from trying to update null UUIDs to non-null. In addition to being not what we want, upstream e2fsck has a bug where it fails to update the checksums when it makes this change, leading to a corrupt superblock.