Fix "Unsubscribe" button on accounts page in Firefox.
Or a less-sensational title that explains the rest of the changeset: avoid
using the global "event" and avoid using "event" as a variable name.

Some browser environments define a global "event" which is set to the current
event that is being dispatched for the duration of event dispatch.  Other
browsers do not.

As a result, since a variable named "event" may or may not shadow the global
event object, it's best to avoid the whole confusion and never name a variable
"event" or use a reference named "event". [1]

[1] - http://eslint.org/docs/rules/no-restricted-globals