Support ext4 journals of non-128M size, and set default to 16M.
It appears that a new e2fsprogs release chooses the default journal size differently, such that the default for an 8G drive has reduced from 128M to 64M. This caused the journal-trimming code to bail out.

This change updates the code to support journals of any size.

Additionally, I've changed the ext4 template to default to a 16M journal. We have had people complain in the past about grains mysteriously using much more storage than advertised. The reason is that the advertised size does not count the journal, but for quota purposes we do count the journal. By reducing the default journal size we can cap the "leaked" storage to 16MB per grain. (Of course, as before, the journal space is recovered when the grain exits.)

In theory this could harm performance of metadata-heavy workloads, but probably Sandstorm apps aren't usually metadata-heavy.