Adjust Mailchimp integration so that if no API key is configured, we only update our local mirror.
This is because I plan to close Sandstorm's Mailchimp account. For our mailing list size, they're charging us $150/month just to store a list. We haven't e-mailed the list in nearly a year. Since we have our own copy of the list in our database, we can easily restore it later -- or maybe implement our own mailing list going forward.